Refactor measures models into separate files.
